requirements.txt
# AI Orchestration
openai>=1.50.0
# The core multi-agent framework
openai-agents>=0.3.0
pydantic>=2.0.0

# Memory (Vector DB)
qdrant-client>=1.12.0

# Systems & Environment
docker>=7.0.0
python-dotenv>=1.0.0

# Utilities
asyncio